+++ src/input.rs
@@ -23,6 +23,7 @@ pub enum Action {
     AutoScrollFaster,
     AutoScrollSlower,
     ToggleIgnoreCase,
+    ToggleMarkdown,
 }
 
 /// Map a key event to an `Action`, decoupled from `AppState` so the keymap
@@ -57,6 +58,7 @@ pub fn map_key(key: KeyEvent) -> Option<Action> {
         (KeyCode::Char('+'), _) | (KeyCode::Char('='), _) => Some(Action::AutoScrollFaster),
         (KeyCode::Char('-'), _) => Some(Action::AutoScrollSlower),
         (KeyCode::Char('i'), _) => Some(Action::ToggleIgnoreCase),
+        (KeyCode::Char('m'), _) => Some(Action::ToggleMarkdown),
         _ => None,
     }
 }